moved gettestkernels.py up into root directory so that the __init__.py file in SpiceyPy is not called.
