+++ gettestkernels.py
